## Purpose

Optimization:
- machine: local
- ensemble=10 parllel=None
- tufte point: run-fire-spread
- Before Optimization: 19.06s, 19.49s, 19.96s
- After Optimization: 15.82s, 16.36s, 15.08s
- Result: 20% improvement

## Related Issues
Closes GRID-157

## Submission Checklist
- [x] Code passes linter

## Testing